use IgnoreLights category to disable WorldUI
usage: WorldUI is used on non-alphablended emissive surfaces to make them emissive. This makes all of the surface emissive. If you want to use a proper emissive mask, add it via a replacement and then assign the ignoreLights category to disable WorldUI. That way, the game still controls the emissive strength